Make project link unicode
